Fix: Debian 11 Bullseye is deprecated, remove it
Debian 11 provided Python 3.9.

This branch removes the support for both Debian 11 and Python 3.9.
The oldest distribution supported is now Ubuntu 22.04 with Python 3.10.

That that mentions of Debian 11 were replaced in
some example files that were not maintained and
the change has not been tested. These remain to
serve as examples for developers.
