+++ .github/workflows/build-deb-package.yml
@@ -9,11 +9,8 @@ jobs:
     strategy:
       fail-fast: false
       matrix:
-        os: ["debian-11", "debian-12", "ubuntu-22.04"]
+        os: ["debian-12", "ubuntu-22.04", "ubuntu-24.04"]
         include:
-          - os: "debian-11"
-            make_target: "all-podman-debian-11"
-            artifact_name: "aleph-vm.debian-11.deb"
           - os: "debian-12"
             make_target: "all-podman-debian-12"
             artifact_name: "aleph-vm.debian-12.deb"
@@ -56,10 +53,8 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        os: ["debian-11", "debian-12"]
+        os: ["debian-12"]
         include:
-          - os: "debian-11"
-            artifact_name: "aleph-debian-11-python.squashfs"
           - os: "debian-12"
             artifact_name: "aleph-debian-12-python.squashfs"
     steps:
